Base Class Update: Adding a minimal working example of the Composition class.
A utils.py has also been added to the library. This holds some functions that are useful for finding the projects paths etc.

A test directory has been created. This should be kept up to date as the project moves on. More advanced numerical tests will be required as the projects grows. The tests should check whether or not modifications have broken anything so run these before you push. I'll try and package the tests into a single python executable to make this faster.