Adding a very basis Stream class and associated test_stream.py. Added heat capacity method to Component class.
We have the ability to create a Stream object and get the mole averaged heat capacity of the stream. Cannot yet add more components after Stream is created. Need a function that runs during Stream.__init__(...) to fill in all the missing data using the properties of the components.

Heat capacity function calls each component heat capacity function in sequence.